Refactor the CLI to work with subcommands, so we can introduce new commands in the future without breaking backwards compatibility then.
