+++ backuppy/tests/cli/test_backuppy.py
@@ -1,21 +1,21 @@
 from unittest import TestCase
 
-from backuppy.cli.backuppy import main
+from backuppy.cli import main
 from backuppy.tests import CONFIGURATION_PATH
 
 
 class CliTest(TestCase):
-    def test_with_json(self):
+    def test_backup_with_json(self):
         configuration_file_path = '%s/backuppy.json' % CONFIGURATION_PATH
-        args = ['-c', configuration_file_path]
+        args = ['backup', '-c', configuration_file_path]
         main(args)
 
-    def test_with_yaml(self):
+    def test_backup_with_yaml(self):
         configuration_file_path = '%s/backuppy.yml' % CONFIGURATION_PATH
-        args = ['-c', configuration_file_path]
+        args = ['backup', '-c', configuration_file_path]
         main(args)
 
-    def test_without_arguments(self):
-        args = []
+    def test_backup_without_arguments(self):
+        args = ['backup']
         with self.assertRaises(SystemExit):
             main(args)
