+++ bin/backuppy
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 import sys
 
-from backuppy.cli.backuppy import main
+from backuppy.cli import main
 
 if __name__ == '__main__':
     main(sys.argv[1:])
